+++ web/static/react/assets/ColdStartInterviewPage-BdwC2FVL.js
@@ -1,0 +1,1 @@
+import{j as r}from"./vendor-react-DCA3L5Xq.js";import{L as t}from"./LegalSkillsProfilePage-CCzZEqux.js";import"./button-DJhBvu1Y.js";import"./index-BvWUaJfh.js";import"./vendor-icons-BX2-bB2Y.js";import"./input-DbaF0qtG.js";import"./IusPageShell-Cvcw1w6o.js";import"./api-DuIksswa.js";function n(){return r.jsx(t,{})}export{n as ColdStartInterviewPage};