+++ web/static/react/assets/PracticeProfilePage-dtGVbhYO.js
@@ -1,1 +1,0 @@
-import{j as r}from"./vendor-react-DCA3L5Xq.js";import{L as o}from"./LegalSkillsProfilePage-C5zgGibK.js";import"./button-CqnlrZV7.js";import"./index-D26seZ1e.js";import"./vendor-icons-BX2-bB2Y.js";import"./input-CzgxJcE0.js";import"./IusPageShell-llX7yOUH.js";import"./api-BJ5BPdiu.js";function l(){return r.jsx(o,{})}export{l as PracticeProfilePage};